angular1seed/server/requirements.txt
Django==1.9.8
django-tastypie==0.13.3
requests==2.10.0
